+++ RELEASE-NOTES.md
@@ -1,5 +1,8 @@
 ## RELEASE NOTES
 ### Version 2.68.0-prerelease-conditional-show-perf"
+**EUI-3055** Fix regression in address field
+
+### Version 2.68.0-prerelease-conditional-show-perf"
 **EUI-3055** Better performance in edit forms through new show hide implementation
 
 ### Version 2.67.4-reset-button-and-keyboard-fix